requirements.txt
django==4.1.5
Pillow==9.4.0
django-admin-sortable2==2.1.4